refactor: 참여자 목록에 중복된 ID 어노테이션 겸증 방식으로 변경
